--- project/src/main/kotlin/edu/isel/ion/core/class_section/ClassSection.kt
@@ -1,3 +1,0 @@
-package edu.isel.ion.core.class_section
-
-class ClassSection
