--- project/src/main/kotlin/edu/isel/ion/core/class_section/ClassSectionController.kt
@@ -1,3 +1,0 @@
-package edu.isel.ion.core.class_section
-
-class ClassSectionController
