--- project/src/main/kotlin/edu/isel/ion/core/class_section/ClassSectionRepo.kt
@@ -1,4 +1,0 @@
-package edu.isel.ion.core.class_section
-
-class ClassSectionRepo {
-}
