--- project/src/main/kotlin/edu/isel/ion/core/class_section/ClassSectionService.kt
@@ -1,2 +1,0 @@
-package edu.isel.ion.core.class_section
-
