--- project/src/main/kotlin/edu/isel/ion/core/course/Course.kt
@@ -1,3 +1,0 @@
-package edu.isel.ion.core.course
-
-class Course
