--- project/src/main/kotlin/edu/isel/ion/core/course/CourseRepo.kt
@@ -1,4 +1,0 @@
-package edu.isel.ion.core.course
-
-class CourseRepo {
-}
