--- project/src/main/kotlin/edu/isel/ion/core/course/CourseService.kt
@@ -1,2 +1,0 @@
-package edu.isel.ion.core.course
-
