--- project/src/main/kotlin/edu/isel/ion/core/course_instance/Class.kt
@@ -1,3 +1,0 @@
-package edu.isel.ion.core.course_instance
-
-class Class
