--- project/src/main/kotlin/edu/isel/ion/core/course_instance/ClassController.kt
@@ -1,3 +1,0 @@
-package edu.isel.ion.core.course_instance
-
-class ClassController
