--- project/src/main/kotlin/edu/isel/ion/core/course_instance/ClassRepo.kt
@@ -1,4 +1,0 @@
-package edu.isel.ion.core.course_instance
-
-class ClassRepo {
-}
