--- project/src/main/kotlin/edu/isel/ion/core/programme/Programme.kt
@@ -1,3 +1,0 @@
-package edu.isel.ion.core.programme
-
-class Programme
