--- project/src/main/kotlin/edu/isel/ion/core/programme/ProgrammeRepo.kt
@@ -1,4 +1,0 @@
-package edu.isel.ion.core.programme
-
-class ProgrammeRepo {
-}
