--- project/src/main/kotlin/edu/isel/ion/core/programme/ProgrammeService.kt
@@ -1,2 +1,0 @@
-package edu.isel.ion.core.programme
-
