+++ project/src/main/kotlin/pt/isel/core/CoreApplication.kt
@@ -1,4 +1,4 @@
-package edu.isel.ion.core
+package pt.isel.core
 
 import org.springframework.boot.autoconfigure.SpringBootApplication
 import org.springframework.boot.runApplication
