+++ project/src/main/kotlin/pt/isel/core/class_section/ClassSection.kt
@@ -1,0 +1,3 @@
+package pt.isel.core.class_section
+
+class ClassSection
