+++ project/src/main/kotlin/pt/isel/core/class_section/ClassSectionRepo.kt
@@ -1,0 +1,4 @@
+package pt.isel.core.class_section
+
+class ClassSectionRepo {
+}
