+++ project/src/main/kotlin/pt/isel/core/class_section/ClassSectionService.kt
@@ -1,0 +1,2 @@
+package pt.isel.core.class_section
+
