+++ project/src/main/kotlin/pt/isel/core/course/Course.kt
@@ -1,0 +1,3 @@
+package pt.isel.core.course
+
+class Course
