+++ project/src/main/kotlin/pt/isel/core/course/CourseController.kt
@@ -1,0 +1,3 @@
+package pt.isel.core.course
+
+class CourseController
