+++ project/src/main/kotlin/pt/isel/core/course/CourseRepo.kt
@@ -1,0 +1,4 @@
+package pt.isel.core.course
+
+class CourseRepo {
+}
