+++ project/src/main/kotlin/pt/isel/core/course_instance/Class.kt
@@ -1,0 +1,3 @@
+package pt.isel.core.course_instance
+
+class Class
