+++ project/src/main/kotlin/pt/isel/core/course_instance/ClassController.kt
@@ -1,0 +1,3 @@
+package pt.isel.core.course_instance
+
+class ClassController
