+++ project/src/main/kotlin/pt/isel/core/course_instance/ClassRepo.kt
@@ -1,0 +1,4 @@
+package pt.isel.core.course_instance
+
+class ClassRepo {
+}
