+++ project/src/main/kotlin/pt/isel/core/course_instance/ClassService.kt
@@ -1,0 +1,2 @@
+package pt.isel.core.course_instance
+
