+++ project/src/main/kotlin/pt/isel/core/programme/Programme.kt
@@ -1,0 +1,3 @@
+package pt.isel.core.programme
+
+class Programme
