+++ project/src/main/kotlin/pt/isel/core/programme/ProgrammeController.kt
@@ -1,0 +1,3 @@
+package pt.isel.core.programme
+
+class ProgrammeController
