+++ project/src/main/kotlin/pt/isel/core/programme/ProgrammeRepo.kt
@@ -1,0 +1,4 @@
+package pt.isel.core.programme
+
+class ProgrammeRepo {
+}
